improved products, partialy implemented shopping cart

+++ src/main/java/com/master/webshop/repositories/ProductRepository.java
@@ -4,6 +4,5 @@ import com.master.webshop.model.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface ProductRepository extends JpaRepository<Product, Integer> {
-
+public interface ProductRepository extends JpaRepository<Product, Long> {
 }
